feat: create an AnnouncementBanner component that composes the Banner
