+++ demo/src/pages/index.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { graphql } from 'gatsby';
 import { css } from '@emotion/core';
 import {
-  Banner,
+  AnnouncementBanner,
   Button,
   CodeBlock,
   Feedback,
@@ -44,9 +44,9 @@ const IndexPage = ({ data }) => {
 
   return (
     <>
-      <Banner>
+      <AnnouncementBanner>
         <p>This is a marketing banner!</p>
-      </Banner>
+      </AnnouncementBanner>
       <GlobalHeader
         editUrl={`${siteMetadata.repository}/tree/develop/demo/src/pages/index.js`}
         search
